Add files via upload
Goal: We are Predicting the Price of gold on the basis of past year data.
We convert the data set of month into a proper date-time object.
Y is continuous column always.
Time Series has   three components :-Trend, Season and Error |Randomness |Residual
Month column will be dropping and convert   into index column of month
Time Series has two types that is Stationary and Non-Stationary time series.
Plot the time series and check  whether  it is stationary time series or not, the time series is not stationary because it has not constant rolling mean that’s the reason time series is in the upward direction.
To get stationary time series, we are using Log technique but we are facing same problem again (getting non-stationary time series)
There is another method where we are taking the difference between log and rolling mean.                 log-Rolling mean
After applying this method we reduces magnitude and ensures that time series in same scale (get constant rolling mean and constant standard deviation)
This time series now making proper forcasting.
For forcating data should be divided into the training and testing 
Here we filter all the null value.
Firstly we create a model and  do training on the basis of Auto Regressive Model.
Then Prediction on the Testing data set.
The Predicted value is against actual value for visualizing.